ACPI / tables: Remap GNVS operation region memory offset in custom DSDTs
We want to override the DSDT table at runtime to fix various issues
caused by mistakes in the DSDT table. However, the DSDT table seems
to differ in the GNVS operation region memory offset on each device.

To allow us to replace the DSDT with a generic fixed version, we
search for the GNVS operation region memory offset in the binary
original table at runtime, and replace the dummy value in the new
table.