ANDROID: drm/i915: Avoid resetting brightness if set to minimum (v2)
Right now, the brightness is reset to maximum if it was set to
minimum before the backlight was disabled. This looks bad when
turning the screen back on because it will actually fade down
instead of fade up when a lower brightness is set.

Just keep the brightness as-is after turning backlight on so
userspace can directly set the correct brightness value.

v2: Keep check but reset brightness only if below minimum.